+++ CHANGES.rst
@@ -1,7 +1,11 @@
 0.0.2 (unreleased)
 -----------------------
 
-- No changes yet
+- Added new functions to download catalogues to
+  `~sdss_catl_utils.mocks_manager.DownloadManager`. It allows for
+  the download of catalogues of specific characteristics.
+- Added new function `~sdss_catl_utils.mocks_manager.CatlUtils` that acts
+  as a tool for handling the synthetic catalogues.
 
 0.0.1 (2018-12-11)
 -----------------------