+++ CHANGES.rst
@@ -4,8 +4,11 @@
 - Added new functions to download catalogues to
   `~sdss_catl_utils.mocks_manager.DownloadManager`. It allows for
   the download of catalogues of specific characteristics.
-- Added new function `~sdss_catl_utils.mocks_manager.CatlUtils` that acts
-  as a tool for handling the synthetic catalogues.
+- Added new function `~sdss_catl_utils.mocks_manager.catl_utils.CatlUtils`
+  that acts as a tool for handling the synthetic catalogues.
+- Added functions `~sdss_catl_utils.mocks_manager.catl_utils.catl_keys` and
+  `~sdss_catl_utils.mocks_manager.catl_utils.catl_keys_prop` to facilitate
+  the handling of the catalogues.
 
 0.0.1 (2018-12-11)
 -----------------------