+++ CHANGES.rst
@@ -9,6 +9,7 @@
 - Added functions `~sdss_catl_utils.mocks_manager.catl_utils.catl_keys` and
   `~sdss_catl_utils.mocks_manager.catl_utils.catl_keys_prop` to facilitate
   the handling of the catalogues.
+- Added unit-testing for `~sdss_catl_utils.mocks_manager.catl_utils` module.
 
 0.0.1 (2018-12-11)
 -----------------------