+++ CHANGES.rst
@@ -10,6 +10,9 @@
   `~sdss_catl_utils.mocks_manager.catl_utils.catl_keys_prop` to facilitate
   the handling of the catalogues.
 - Added unit-testing for `~sdss_catl_utils.mocks_manager.catl_utils` module.
+- Added more functions for `~sdss_catl_utils.mocks_manager.catl_utils` module
+  (`~sdss_catl_utils.mocks_manager.catl_utils.catl_prefix_main`,
+  `~sdss_catl_utils.mocks_manager.catl_utils.check_input_params`)
 
 0.0.1 (2018-12-11)
 -----------------------